@@ -1,23 +1,22 @@
 {
-    "name": "ambta/doctrine-encrypt-bundle",
+    "name": "integralservice/doctrine-encrypt-bundle",
     "type": "library",
-    "keywords": ["doctrine", "symfony", "aes256", "rijndael", "encrypt", "decrypt"],
+    "keywords": ["doctrine", "symfony", "encrypt", "decrypt"],
     "license": "MIT",
     "description": "Symfony 2 bundle which allows to encrypt data in database with some encrypt algorithm",
     "require": {
-        "php": ">=5.4",
-        "symfony/framework-bundle": ">=2.5",
-        "doctrine/orm": ">=2.5",
-        "ext-mcrypt": "*"
+        "php": ">=7.2",
+        "symfony/framework-bundle": ">=3.4",
+        "doctrine/orm": ">=2.5"
     },
     "autoload": {
-        "psr-0": { "Ambta\\DoctrineEncryptBundle": "" }
+        "psr-0": { "IntegralService\\DoctrineEncryptBundle": "" }
     },
-    "target-dir": "Ambta/DoctrineEncryptBundle",
+    "target-dir": "IntegralService/DoctrineEncryptBundle",
     "authors": [
         {
-            "name": "Marcel van Nuil",
-            "email": "marcel@ambta.com"
+            "name": "Julien Rouvier",
+            "email": "julien@integral-service.fr"
         }
     ]
 }
